Move conversion preprocessor tests into test_preprocess
